backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "numpy>=2.4.4",
    "pydantic-settings>=2.14.0",
    "pytest>=9.0.3",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.45.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "polyfactory>=3.3.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
